challenge_1/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
pgvector==0.2.4

# Caching
redis==5.0.1
hiredis==2.2.3

# AI/ML and embeddings
openai==1.3.7
google-generativeai==0.3.2
sentence-transformers==2.2.2
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0
numpy==1.24.4
scipy==1.11.4

# Text processing and NLP
spacy==3.7.2
nltk==3.8.1
markdown==3.5.1
python-docx==1.1.0
PyPDF2==3.0.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Async support
aiofiles==23.2.1
aioredis==2.0.1
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# File handling
magic==0.4.27
filetype==1.2.0

# Performance
orjson==3.9.10

# Environment and configuration
python-decouple==3.8